fix Dockerfile and rename app
