+++ buho_back/config.py
@@ -1,4 +1,3 @@
-# app/config.py
 import os
 import json
 from pydantic_settings import BaseSettings
